Move metadata, change references
